% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_prob_thief.R
\name{covid_prob_thief}
\alias{covid_prob_thief}
\title{Temporal hierarchical forecasting for COVID-19 incident hospitalizations for
multiple locations}
\usage{
covid_prob_thief(
  truth_data = NULL,
  ts_col = "value",
  start_date,
  end_date,
  fips_vec,
  target_name,
  aggregate_levels,
  frequency,
  nsim = 1e+05,
  n_samples = NULL,
  quantile_levels,
  transform.4root = FALSE
)
}
\arguments{
\item{truth_data}{A data frame containing truth data used to create forecasts
Must contain a "target_end_date" column of dates, a "location" column of
fips codes, and a column of truth data values. Defaults to NULL, in
which case hospitalization truth data is sourced as of the user-specified
\code{end_date}.}

\item{ts_col}{The name of the column containing the truth data. This column
is coerced into a time series object of class \code{ts} and thus should
be a numeric type.}

\item{start_date}{A date from which the truth data begins.}

\item{end_date}{A date where the truth data ends and the forecasts begin.
If \code{df = NULL}, also specifies the date from which the hospitalization
truth data sourced.}

\item{fips_vec}{A vector of one or more 2-digit codes specifying a
United States state or territory, of class \code{char}.}

\item{aggregate_levels}{A user-selected list of aggregates to use.}

\item{frequency}{Integer seasonal period.}

\item{nsim}{Numeric of bootstrap samples used to generate probabilistic
forecasts. Defaults to 10000.}

\item{n_samples}{Numeric giving the number of samples for each unique
forecast unit to return. Defaults to NULL, in which case no
sample forecasts are returned.}

\item{quantile_levels}{Numeric vector of quantile levels (probabilities) to
calculate for the returned forecasts.}

\item{transform.4root}{\code{logical} that specifies whether a variance
stabilizing fourth root transformation should be performed on the data.
(This data transformation is undone after all of the forecasts are
reconciled and re-formatted into a data frame.)}
}
\value{
A list containing two data frames: the first containing COVID-19
incident hospitalization forecasts for the specified date and locations
in a US COVID-19 Forecast Hub format and the second containing metadata
about the forecaster making the predictions for each location, with
columns "forecast_data", "location", aggregate_levels, "base_matrix",
"rec_matrix" (the matrices contain the original sample forecasts)
}
\description{
Temporal hierarchical forecasting for COVID-19 incident hospitalizations for
multiple locations
}
