% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thief.R
\name{get_ts_dates}
\alias{get_ts_dates}
\title{Get a vector of the actual dates for plotting}
\usage{
get_ts_dates(start_date, most_recent_date, frequency = 56)
}
\arguments{
\item{start_date}{A date from which the data begins. Used to calculate the
actual dates of the time series.}

\item{most_recent_date}{A date where the truth data ends. Used to calculate
the actual dates of the time series.}

\item{frequency}{Integer seasonal period. Specifies how many actual dates
to calculate.}
}
\value{
A vector of dates of length \code{frequency}.
}
\description{
Get a vector of the actual dates for plotting
}
