% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prob_thief.R
\name{intervals_to_quantiles}
\alias{intervals_to_quantiles}
\title{Obtain the associated quantile levels from provided interval levels}
\usage{
intervals_to_quantiles(pi_levels)
}
\arguments{
\item{pi_levels}{A numeric vector of interval levels to transform to
quantile levels. Must be an integer between 1 and 100.}
}
\value{
A numeric vector of quantile levels (probabilities)
}
\description{
Obtain the associated quantile levels from provided interval levels
}
