% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wis_loc.R
\name{next_monday}
\alias{next_monday}
\title{All functions in script adapted from Cramer, et al
Source: \url{https://github.com/reichlab/covid19-forecast-evals/blob/main/code/figure-wis_by_location.R}}
\usage{
next_monday(date)
}
\description{
All functions in script adapted from Cramer, et al
Source: \url{https://github.com/reichlab/covid19-forecast-evals/blob/main/code/figure-wis_by_location.R}
}
