% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prob_thief.R
\name{plot_prob_thief}
\alias{plot_prob_thief}
\title{Plot probabilistic reconciled and base point temporal hierarchical forecasts
for every level of the hierarchy}
\usage{
plot_prob_thief(base_matrix, reco_matrix, forecast_date, truth_data = NULL)
}
\arguments{
\item{base_matrix}{A \code{matrix} containing base forecasts to be plotted.
The assumed matrix format contains columns of samples for each element in
the hierarchy, with one sample per row.}

\item{reco_matrix}{A \code{matrix} containing reconciled forecasts to be
plotted. The assumed matrix format contains columns of samples for
each element in the hierarchy, with one sample per row.}

\item{forecast_date}{A \code{Date} from which the forecasts are relevant to.
Used to calculate the dates for x-axis labels.}

\item{truth_data}{A data frame containing truth data used to create forecasts
Must contain a "target_end_date" column of dates, a "location" column of
fips codes, and a column of truth data values. Defaults to NULL, in
which case hospitalization truth data is sourced as of the user-specified
\code{end_date}.}

\item{aggregate_truth_data}{List of (hierarchical) time series containing
truth data to be plotted against the forecasts for comparison. Can be the
input to the \code{temporal_hierarchy} parameter used to calculate forecasts,
OR an extension of the truth data through the last target end date for the
forecasts. Defaults to NULL in which no truth data line is shown.}

\item{quantile_levels}{Numeric vector of quantile levels (probabilities) to
be plotted for the provided \code{base_forecast} and `reconciled_forecasts.
Defaults to 50\% and 95\% intervals.}
}
\value{
A collection of plots made at every aggregation level displaying base
point forecasts, reconciled point forecasts, reconciled 50\% and 95\%
prediction intervals, and the truth data used to make the forecasts
}
\description{
Plot probabilistic reconciled and base point temporal hierarchical forecasts
for every level of the hierarchy
}
