% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summarized_metrics.R
\name{plot_summarized_metrics}
\alias{plot_summarized_metrics}
\title{Plot summarized metrics against horizon week}
\usage{
plot_summarized_metrics(
  summarized_metrics,
  model_names,
  model_colors,
  y_var = "WIS",
  main
)
}
\arguments{
\item{summarized_metrics}{A data frame of summarized metrics. Must contain one row per model and horizon week combination plus a \code{horizon_wk} column}

\item{model_names}{An ordered vector of model names}

\item{model_colors}{An ordered vector of model colors. Must match with \code{model_names} order}

\item{y_var}{A string specifying which metric to plot as the y-variable}

\item{main}{A string specifying the plot title}
}
\value{
A scatter plot (with observations connected by lines) of the specified summary metric vs horizon week
}
\description{
Plot summarized metrics against horizon week
}
