% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thief_agg.R
\name{plot_thief_agg}
\alias{plot_thief_agg}
\title{Plot aggregated time series data}
\usage{
plot_thief_agg(temporal_hierarchy, start_date)
}
\arguments{
\item{temporal_hierarchy}{A list of (hierarchical) time series.}

\item{start_date}{A date from which the data begins. Used to calculate date labels for the plot.}
}
\value{
A plot of the input time series stacked on top of each other, sharing an x-axis.
}
\description{
Plot aggregated time series data
}
