% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wis_loc.R
\name{plot_wis_loc}
\alias{plot_wis_loc}
\title{Calculate and plot pairwise WIS by location}
\usage{
plot_wis_loc(scores, truth, model_levels, baseline_name)
}
\arguments{
\item{scores}{A data frame of forecast scores to plot}

\item{truth}{A data frame of truth data}

\item{model_levels}{Ordered vector of model names used as axis levels}

\item{baseline_name}{Name of baseline model (must be included in model_levels)}
}
\value{
A tile plot of pairwise relative WIS broken down by location and model using the scores provided
}
\description{
Calculate and plot pairwise WIS by location
}
