% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_thief_wrapper.R
\name{prob_thief_wrapper}
\alias{prob_thief_wrapper}
\title{Probabilistic temporal hierarchical forecasting for a single location}
\usage{
prob_thief_wrapper(
  truth_data = NULL,
  ts_col = "value",
  start_date,
  end_date,
  fips_code,
  target_name = "inc hosp",
  aggregate_levels,
  frequency,
  nsim = 1e+05,
  n_samples = NULL,
  quantile_levels = NULL,
  transform.4root = FALSE
)
}
\arguments{
\item{truth_data}{A data frame containing truth data used to create forecasts
Must contain a "target_end_date" column of dates, a "location" column of
fips codes, and a column of truth data values. Defaults to NULL, in
which case hospitalization truth data is sourced as of the user-specified
\code{end_date}.}

\item{ts_col}{The name of the column containing the truth data. This column
is coerced into a time series object of class \code{ts} and thus should be
a numeric type.}

\item{start_date}{A date from which the truth data begins.}

\item{end_date}{A date where the truth data ends and the forecasts begin.
If \code{df = NULL}, also specifies the date from which the hospitalization
truth data sourced.}

\item{fips_code}{A 2-digit code specifying a United States state or territory
of type \code{char}.}

\item{target_name}{A character string giving the name to use for the target.}

\item{aggregate_levels}{A user-selected list of aggregates to use.}

\item{frequency}{Integer seasonal period.}

\item{nsim}{Numeric of bootstrap samples used to generate probabilistic
forecasts. Defaults to 10000.}

\item{n_samples}{Numeric giving the number of samples for each unique
forecast unit to return. Defaults to NULL, in which case no
sample forecasts are returned.}

\item{quantile_levels}{Numeric vector of quantile levels (probabilities) to
calculate for the returned forecasts.}

\item{transform.4root}{\code{logical} that specifies whether a variance
stabilizing fourth root transformation should be performed on the data.
(This data transformation is undone after all of the forecasts are
reconciled and re-formatted into a data frame.)}
}
\value{
A list containing two items: a data frame containing COVID-19
incident hospitalization truth data with a US COVID-19 Forecast Hub format
and a data frame containing the base forecast matrix and the reconciled
forecast matrix with other relevant identifying information.
}
\description{
Probabilistic temporal hierarchical forecasting for a single location
}
