% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_forecast_metrics.R
\name{summarize_forecast_date_metrics}
\alias{summarize_forecast_date_metrics}
\title{Summarize forecasts by forecast date and horizon week}
\usage{
summarize_forecast_date_metrics(scores, us_only)
}
\arguments{
\item{scores}{A data frame of forecast scores to be summarized}

\item{us_only}{Boolean specifying whether to summarize metrics for the us national level only or just states}
}
\value{
A data frame of summarized forecast score metrics stratified by forecast date and horizon week
}
\description{
Summarize forecasts by forecast date and horizon week
}
