% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_forecast_metrics.R
\name{summarize_horizon_metrics}
\alias{summarize_horizon_metrics}
\title{Summarize forecasts by horizon week}
\usage{
summarize_horizon_metrics(scores, baseline_name, us_only = FALSE)
}
\arguments{
\item{scores}{A data frame of forecast scores to be summarized}

\item{baseline_name}{String specifying the name of the baseline model to
calculate relative metrics against}

\item{us_only}{Boolean specifying whether to summarize metrics for the
US national level only or just states}
}
\value{
A data frame of summarized forecast score metrics stratified by horizon week
}
\description{
Summarize forecasts by horizon week
}
