% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_matrix_to_hub_df.R
\name{transform_matrix_to_hub_df}
\alias{transform_matrix_to_hub_df}
\title{Transform temporal hierarchical COVID-19 probabilistic forecasts in a matrix
to US COVID-19 Forecast Hub-formatted data frame of quantile forecasts}
\usage{
transform_matrix_to_hub_df(
  fc_matrix,
  forecast_date,
  fips_code,
  target_name,
  n_samples = NULL,
  quantile_levels = c(0.025, 0.25, 0.5, 0.75, 0.975),
  h_ahead = 56,
  temp_res = 1,
  keep_bottommost_only = TRUE
)
}
\arguments{
\item{fc_matrix}{Object of class \code{matrix} containing forecasts to be
transformed into a data frame. The assumed matrix format contains columns
of samples for each element in the hierarchy, with one sample per row.
The columns are named with format "k-agg h-element", e.g. "k-28 h-1"
represents the 28-day level's first element.}

\item{forecast_date}{Date from which horizons begin counting from (h = 0),
tends to be the last observed value in the truth data.}

\item{fips_code}{A 2-digit code specifying a United States state or territory
of type \code{char}. Used to set the \code{location} in the new data frame.}

\item{target_name}{A character string giving the name to use for the target.}

\item{n_samples}{Numeric giving the number of samples for each unique
forecast unit to return. Defaults to NULL, in which case no
sample forecasts are returned.}

\item{quantile_levels}{Numeric vector of quantile levels (probabilities) to
calculate from the input \code{fc_matrix} sample forecasts. NULL means that no
quantile forecasts are returned. Defaults to quantiles representing
50\% and 95\% prediction intervals and a median.}

\item{h_ahead}{Numeric specifying the time units ahead of the longest horizon
being forecast should be. Defaults to 56 with the assumption the unit is
in days (for a total of 8 weeks ahead).}

\item{keep_bottommost_only}{Logical specifying whether to discard all
forecasts except for those from the bottommost level of the hierarchy.
Defaults to TRUE.}
}
\value{
A data frame of forecasts containing columns: \code{forecast_date},
\code{location}, \code{target}, \code{target_end_date}, \code{type},
\code{quantile}, \code{value}. Quantile values are summarized from the input
sample forecasts.
}
\description{
Transform temporal hierarchical COVID-19 probabilistic forecasts in a matrix
to US COVID-19 Forecast Hub-formatted data frame of quantile forecasts
}
