% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_thief_df.R
\name{aggregate_thief_df}
\alias{aggregate_thief_df}
\title{Non-overlapping temporal aggregation of COVID-19 truth data stored as a list}
\usage{
aggregate_thief_df(
  df = NULL,
  ts_col = "value",
  start_date,
  end_date,
  fips_code,
  aggregate_levels = list(56, 28, 14, 7, 1),
  agg.names = NULL,
  frequency = 56,
  transform.4root = FALSE
)
}
\arguments{
\item{df}{A data frame containing the desired truth data as one of the
columns. Used to create the forecasts. Defaults to NULL in which
hospitalization truth data is sourced as of the user-specified
\code{end_date}. Must have "target_end_date" and "location" as columns.}

\item{ts_col}{The name of the column containing the truth data. This column
is coerced into a time series object of class \code{ts} and thus should
be a numeric type.}

\item{start_date}{A date from which the truth data begins.}

\item{end_date}{A date where the truth data ends and the forecasts begin.
If \code{df = NULL}, also specifies the date from which the hospitalization
truth data sourced.}

\item{fips_code}{A 2-digit code specifying a United States state or territory
of type \code{char}.}

\item{aggregate_levels}{A user-selected list of aggregates to use.}

\item{agg.names}{A vector of names for the aggregation levels. Will be used
for any plots generated later split by aggregation level. Defaults to NULL.}

\item{frequency}{Integer seasonal period.}

\item{transform.4root}{\code{logical} that specifies whether a variance
stabilizing fourth root transformation should be performed on the data.}
}
\value{
A list of time series. The first element is a \code{ts} object
made by coercing the original truth data into a time series format,
followed by series with increasing levels of aggregation.
}
\description{
Non-overlapping temporal aggregation of COVID-19 truth data stored as a list
}
