% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_base_forecasts.R
\name{compute_base_forecasts}
\alias{compute_base_forecasts}
\title{Compute temporal hierarchical (base) forecasts}
\usage{
compute_base_forecasts(temporal_hierarchy, pi_levels = c(10 * (1:9), 95, 98))
}
\arguments{
\item{temporal_hierarchy}{A list of (hierarchical) time series.}

\item{pi_levels}{A vector of prediction interval levels to calculate. Defaults to the levels that correspond to the 23 US Forecast Hub-specified quantiles.}
}
\value{
An object of class \code{forecast}.
}
\description{
Compute temporal hierarchical (base) forecasts
}
