% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prob_forecasts.R
\name{compute_prob_forecasts}
\alias{compute_prob_forecasts}
\title{Compute temporal hierarchical probabilistic forecasts}
\usage{
compute_prob_forecasts(
  temporal_hierarchy,
  nsim = 10000,
  forecast_type = c("base", "reconciled"),
  aggregate_levels,
  ...
)
}
\arguments{
\item{temporal_hierarchy}{List of (hierarchical) time series.}

\item{nsim}{Numeric of bootstrap samples used to generate probabilistic
forecasts. Defaults to 10000.}

\item{forecast_type}{Character string or vector specifying the types of
forecasts to return. May be "base", "reconciled", or both.}

\item{aggregate_levels}{Numeric vector of aggregation levels, where each
value is described relative to the smallest unit. For example, c(14, 7, 1)
can represent a hierarchy with levels 2-weekly, weekly, and daily. Only
required when \code{forecast_type} contains "reconciled"}

\item{...}{Additional arguments passed onto \code{FoReco::terec}}
}
\value{
Either a matrix or a list of matrices (depending on the number of
forecast types requested) containing probabilistic temporal hierarchical
forecasts represented by samples. A single matrix has columns representing
each element in the provided hierarchy, with one sample per row.
}
\description{
Compute temporal hierarchical probabilistic forecasts
}
