% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidTHieF-package.R
\docType{package}
\name{covidTHieF-package}
\alias{covidTHieF}
\alias{covidTHieF-package}
\title{covidTHieF: Covid-19 Temporal Hierarchical Forecasting}
\description{
Repository containing the code and paper on forecasting COVID-19 hospitalizations using the THieF method (Athanasopoulos et al. 2017 \doi{10.1016/j.ejor.2017.02.046}) and its extension FoReco (Girolimetto et al. 2023 \doi{10.1016/j.ijforecast.2023.10.003})
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hubverse-org/hubEnsemblesManuscript}
}

}
\author{
\strong{Maintainer}: Li Shandross \email{lshandross@umass.edu} (\href{https://orcid.org/0009-0008-1348-1954}{ORCID})

}
\keyword{internal}
