% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_formatted_forecasts.R
\name{load_formatted_forecasts}
\alias{load_formatted_forecasts}
\title{Load available forecasts from local directory (optimized for parallelization)}
\usage{
load_formatted_forecasts(model_vector, date_indices)
}
\arguments{
\item{model_vector}{the name of a single model}

\item{date_indices}{a vector of numerics corresponding to associated forecasts dates}
}
\value{
a data frame containing the forecasts from a single model during the
corresponding dates to the specified date indices
}
\description{
Load available forecasts from local directory (optimized for parallelization)
}
