% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summarized_metrics.R
\name{plot_forecast_date_metrics}
\alias{plot_forecast_date_metrics}
\title{Plot summarized metrics against forecast date}
\usage{
plot_forecast_date_metrics(
  forecast_date_metrics,
  model_names,
  model_colors,
  y_var = c("WIS", "MAE", "Cov95", "Cov50"),
  horizon_week,
  main
)
}
\arguments{
\item{forecast_date_metrics}{A data frame of forecast_date metrics. Must contain one row per model and horizon week combination plus a \code{horizon_wk} column}

\item{model_names}{An ordered vector of model names}

\item{model_colors}{An ordered vector of model colors. Must match with \code{model_names} order}

\item{y_var}{A string specifying which metric to plot as the y-variable. Must be one of "WIS", "MAE", "Cov95", "Cov50"}

\item{horizon_week}{An integer specifying the horizon week of the desired metric to plot}

\item{main}{A string specifying the plot title}
}
\value{
A scatter plot (with observations connected by lines) of the specified summary metric vs forecast date
}
\description{
Plot summarized metrics against forecast date
}
