% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_models.R
\name{plot_models_one_location}
\alias{plot_models_one_location}
\title{Plot the forecasts of multiple models for a single location}
\usage{
plot_models_one_location(
  forecasts,
  truth,
  models = NULL,
  fips = "US",
  fc_dates = NULL,
  prediction_intervals = c(0.5, 0.95),
  facet_nrow = 6,
  date_limits = NULL
)
}
\arguments{
\item{forecasts}{A data frame of forecasts to plot}

\item{truth}{A data frame of truth data}

\item{fips}{A string of the fips code for the desired location to plot}

\item{fc_dates}{A vector of the forecast dates to plot. Defaults to NULL where
all provided forecasts will be plotted}

\item{prediction_intervals}{A vector of (1-alpha)*100\% prediction intervals to
plot in addition to point forecasts. Defaults to 50\% and 95\%.}

\item{facet_nrow}{An integer specifying how many rows the grid of forecast plots
should have. Defaults to 6}

\item{date_limits}{A vector containing two dates between which to restrict the
x-axis (in case of extra long truth data)}
}
\value{
A plot of provided forecasts and truth data, faceted by model, for a single location
}
\description{
Plot the forecasts of multiple models for a single location
}
