% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thief.R
\name{plot_thief}
\alias{plot_thief}
\title{Plot base and reconciled temporal hierarchical forecasts.}
\usage{
plot_thief(
  base_forecasts,
  reconciled_forecasts,
  ts_dates,
  extended_truth = NULL,
  agg.names = NULL
)
}
\arguments{
\item{base_forecasts}{A \code{forecast} object containing the original
(unreconciled) base forecasts.}

\item{reconciled_forecasts}{A \code{forecast} object containing reconciled
forecasts. If prediction intervals are specified, two will be shown.}

\item{ts_dates}{A vector of dates to be used as x-axis labels.}

\item{extended_truth}{A list of time series to be plotted against the
forecasts for comparison. Defaults to NULL in which no extended truth line
is shown.}

\item{agg.names}{A vector of titles for the plots of each aggregation level.
Defaults to NULL in which the detected ARIMA model is used to name the
plots.}
}
\value{
A collection of plots made at every aggregation level displaying base
point forecasts, reconciled point forecasts, reconciled prediction
intervals at two different levels, and the truth data used to make the
forecasts. The user may provide extended truth data to plot against the
forecasts for an accuracy comparison.
}
\description{
Plot base and reconciled temporal hierarchical forecasts.
}
