% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_hub_df.R
\name{transform_to_hub_df}
\alias{transform_to_hub_df}
\title{Transform temporal hierarchical COVID-19 forecasts to US COVID-19 Forecast
Hub-formatted data frame}
\usage{
transform_to_hub_df(
  forecasts,
  most_recent_date,
  fips_code,
  pi_levels,
  h_ahead = 56,
  transform.4root = FALSE
)
}
\arguments{
\item{forecasts}{An object of class \code{forecast} to be transformed into a
data frame}

\item{most_recent_date}{A date from which the truth data ends and the day
before the forecasts begin. Used to set the \code{forecast_date} in the
new data frame.}

\item{fips_code}{A 2-digit code specifying a United States state or territory
of type \code{char}. Used to set the \code{location} in the new data frame.}

\item{pi_levels}{A vector of prediction interval levels to calculate. Used
to obtain the corresponding \code{quantile} value in the new data frame.}

\item{h_ahead}{A number specifying how many time units ahead the longest
horizon being forecast should be. Defaults to 56 with the assumption the
unit is in days (for a total of 8 weeks ahead).}

\item{transform.4root}{\code{logical} that specifies whether a variance
stabilizing fourth root transformation was performed on the data when
creating the provided forecasts. If \code{TRUE}, the forecast values are
raised to the fourth power to undo the initial transformation.}
}
\value{
A data frame containing the following columns: \code{forecast_date},
\code{location}, \code{target}, \code{target_end_date}, \code{type},
\code{quantile}, \code{value}
}
\description{
Transform temporal hierarchical COVID-19 forecasts to US COVID-19 Forecast
Hub-formatted data frame
}
